Shifted views around. Actually locked down read-only permissions. Somehow this broke no tests.

+++ lib/mpnetwork_web/controllers/broadcast_controller.ex
@@ -1,7 +1,7 @@
 defmodule MpnetworkWeb.BroadcastController do
   use MpnetworkWeb, :controller
 
-  alias Mpnetwork.Realtor
+  alias Mpnetwork.{Realtor, Permissions}
 
   @index_max 20
 
@@ -11,20 +11,28 @@ defmodule MpnetworkWeb.BroadcastController do
   end
 
   def new(conn, _params) do
-    changeset = Realtor.change_broadcast(%Realtor.Broadcast{})
-    render(conn, "new.html", changeset: changeset)
+    if !Permissions.read_only?(current_user(conn)) do
+      changeset = Realtor.change_broadcast(%Realtor.Broadcast{})
+      render(conn, "new.html", changeset: changeset)
+    else
+      send_resp(conn, 405, "Not allowed")
+    end
   end
 
   def create(conn, %{"broadcast" => broadcast_params}) do
     # inject current_user.id
-    broadcast_params_with_current_user_id = Enum.into(%{"user_id" => current_user(conn).id}, broadcast_params)
-    case Realtor.create_broadcast(broadcast_params_with_current_user_id) do
-      {:ok, broadcast} ->
-        conn
-        |> put_flash(:info, "Broadcast created successfully.")
-        |> redirect(to: broadcast_path(conn, :show, broadcast))
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, "new.html", changeset: changeset)
+    if !Permissions.read_only?(current_user(conn)) do
+      broadcast_params_with_current_user_id = Enum.into(%{"user_id" => current_user(conn).id}, broadcast_params)
+      case Realtor.create_broadcast(broadcast_params_with_current_user_id) do
+        {:ok, broadcast} ->
+          conn
+          |> put_flash(:info, "Broadcast created successfully.")
+          |> redirect(to: broadcast_path(conn, :show, broadcast))
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, "new.html", changeset: changeset)
+      end
+    else
+      send_resp(conn, 405, "Not allowed")
     end
   end
 
@@ -34,27 +42,35 @@ defmodule MpnetworkWeb.BroadcastController do
   end
 
   def edit(conn, %{"id" => id}) do
-    broadcast = Realtor.get_broadcast!(id)
-    if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
-      changeset = Realtor.change_broadcast(broadcast)
-      render(conn, "edit.html", broadcast: broadcast, changeset: changeset)
+    if !Permissions.read_only?(current_user(conn)) do
+      broadcast = Realtor.get_broadcast!(id)
+      if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
+        changeset = Realtor.change_broadcast(broadcast)
+        render(conn, "edit.html", broadcast: broadcast, changeset: changeset)
+      else
+        send_resp(conn, 405, "Not allowed")
+      end
     else
       send_resp(conn, 405, "Not allowed")
     end
   end
 
   def update(conn, %{"id" => id, "broadcast" => broadcast_params}) do
-    broadcast = Realtor.get_broadcast!(id)
-    # inject current_user.id
-    broadcast_params_with_current_user_id = Enum.into(%{"user_id" => current_user(conn).id}, broadcast_params)
-    if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
-      case Realtor.update_broadcast(broadcast, broadcast_params_with_current_user_id) do
-        {:ok, broadcast} ->
-          conn
-          |> put_flash(:info, "Broadcast updated successfully.")
-          |> redirect(to: broadcast_path(conn, :show, broadcast))
-        {:error, %Ecto.Changeset{} = changeset} ->
-          render(conn, "edit.html", broadcast: broadcast, changeset: changeset)
+    if !Permissions.read_only?(current_user(conn)) do
+      broadcast = Realtor.get_broadcast!(id)
+      # inject current_user.id
+      broadcast_params_with_current_user_id = Enum.into(%{"user_id" => current_user(conn).id}, broadcast_params)
+      if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
+        case Realtor.update_broadcast(broadcast, broadcast_params_with_current_user_id) do
+          {:ok, broadcast} ->
+            conn
+            |> put_flash(:info, "Broadcast updated successfully.")
+            |> redirect(to: broadcast_path(conn, :show, broadcast))
+          {:error, %Ecto.Changeset{} = changeset} ->
+            render(conn, "edit.html", broadcast: broadcast, changeset: changeset)
+        end
+      else
+        send_resp(conn, 405, "Not allowed")
       end
     else
       send_resp(conn, 405, "Not allowed")
@@ -62,13 +78,17 @@ defmodule MpnetworkWeb.BroadcastController do
   end
 
   def delete(conn, %{"id" => id}) do
-    broadcast = Realtor.get_broadcast!(id)
-    if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
-      {:ok, _broadcast} = Realtor.delete_broadcast(broadcast)
+    if !Permissions.read_only?(current_user(conn)) do
+      broadcast = Realtor.get_broadcast!(id)
+      if current_user(conn).id == broadcast.user_id || current_user(conn).role_id < 3 do
+        {:ok, _broadcast} = Realtor.delete_broadcast(broadcast)
 
-      conn
-      |> put_flash(:info, "Broadcast deleted successfully.")
-      |> redirect(to: broadcast_path(conn, :index))
+        conn
+        |> put_flash(:info, "Broadcast deleted successfully.")
+        |> redirect(to: broadcast_path(conn, :index))
+      else
+        send_resp(conn, 405, "Not allowed")
+      end
     else
       send_resp(conn, 405, "Not allowed")
     end
